+++ Menu.hpp
@@ -22,7 +22,7 @@ namespace menu {
      * @return the value of type T entered by the user.
      */
     template<typename T>
-    T prompt(const std::string &prompt, const int indent_val = 0) {
+    T prompt(const std::string &prompt, const size_t indent_val = 0) {
         T choice;
         std::string indent(indent_val, ' ');
         std::cout << indent << prompt << "? ";
@@ -50,7 +50,7 @@ namespace menu {
      * @return the value of type T entered by the user that is within the specified range
      */
     template<typename T>
-    T prompt_range(const std::string &prompt, const T &minValid, const T &maxValid, const int indent_val = 0) {
+    T prompt_range(const std::string &prompt, const T &minValid, const T &maxValid, const size_t indent_val = 0) {
         T choice;
         std::string indent(indent_val, ' ');
         std::cout << indent << prompt << " (" << minValid << ", " << maxValid << ")? ";
@@ -74,7 +74,7 @@ namespace menu {
      * @param indent_val the amount of spaces the prompt should be indented
      * @return the user's choice between 1 and options.size() (inclusive)
      */
-    int prompt_option(const std::string &prompt, const std::vector<std::string>& options, int indent_val = 0);
+    size_t prompt_option(const std::string &prompt, const std::vector<std::string>& options, size_t indent_val = 0);
 
     class MenuItem {
     private:
@@ -127,7 +127,7 @@ namespace menu {
 
     class Menu {
     private:
-        int indent{};
+        size_t indent{};
         std::vector<MenuItem> items;
 
     public:
@@ -164,7 +164,7 @@ namespace menu {
          *
          * @return the amount of spaces this Menu Should be indented
          */
-        int getIndent();
+        size_t getIndent();
 
         /*
          * run this menu. The menu will prompt the user with a list
